separate wallet from types
